Rename assistant_state
