+++ lib/ui/notifications.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:global_configuration/global_configuration.dart';
 import 'package:summer2022/main.dart';
 import 'package:summer2022/ui/top_app_bar.dart';
-import 'AssistantState.dart';
+import 'assistant_state.dart';
 import 'bottom_app_bar.dart';
