+++ lib/ui/sign_in.dart
@@ -8,7 +8,7 @@ import 'package:summer2022/services/analytics_service.dart';
 import 'package:summer2022/services/cache_service.dart';
 import 'package:summer2022/utility/locator.dart';
 
-import 'AssistantState.dart';
+import 'assistant_state.dart';
 
 class SignInWidget extends StatefulWidget {
   const SignInWidget({Key? key}) : super(key: key);
